Add files via upload
Problem Statement:
You are the data scientist at a telecom company named “Neo” whose customers
are churning out to its competitors. You have to analyze the data of your
company and find insights and stop your customers from churning out to other
telecom companies.
Tasks To Be Performed:

1. Data Manipulation:
● Extract the 5th column and store it in ‘customer_5’
● Extract the 15th column and store it in ‘customer_15’
● Extract all the male senior citizens whose payment method is electronic
check and store the result in ‘senior_male_electronic’
● Extract all those customers whose tenure is greater than 70 months or
their monthly charges is more than $100 and store the result in
‘customer_total_tenure’
Extract all the customers whose contract is of two years, payment method
is mailed check and the value of churn is ‘Yes’ and store the result in
‘two_mail_yes’
● Extract 333 random records from the customer_churndataframe and store
the result in ‘customer_333’
● Get the count of different levels from the ‘Churn’ column

2. Data Visualization:
● Build a bar-plot for the ’InternetService’ column:
a. Set x-axis label to ‘Categories of Internet Service’
b. Set y-axis label to ‘Count of Categories’
c. Set the title of plot to be ‘Distribution of Internet Service’
d. Set the color of the bars to be 'orange'
Build a histogram for the ‘tenure’ column:
a. Set the number of bins to be 30
b. Set the color of the bins to be ‘green’
c. Assign the title 'distribution'